+++ packages/browser-core/src/browser/pageMayExitObservable.ts
@@ -1,10 +1,10 @@
-import { globalObject, objectValues } from '@datadog/js-core/util'
+import { globalObject } from '@datadog/js-core/util'
 import { PageExitReason } from '@datadog/js-core/transport'
 import type { PageMayExitEvent } from '@datadog/js-core/transport'
 import { Observable } from '../tools/observable'
 import { addEventListeners, addEventListener, DOM_EVENT } from './addEventListener'
 
-export { PageExitReason } from '@datadog/js-core/transport'
+export { PageExitReason, isPageExitReason } from '@datadog/js-core/transport'
 export type { PageMayExitEvent } from '@datadog/js-core/transport'
 
 export function createPageMayExitObservable(): Observable<PageMayExitEvent> {
@@ -46,6 +46,4 @@ export function createPageMayExitObservable(): Observable<PageMayExitEvent> {
   })
 }
 
-export function isPageExitReason(reason: string): reason is PageExitReason {
-  return objectValues(PageExitReason).includes(reason as PageExitReason)
-}
+
